src/requirements.txt
opencv-python
numpy==1.26.1
pandas==1.5.3
tensorflow==2.14.0
seaborn==0.12.2
matplotlib.pyplot